Add Serbian as a common pluralizations file
